modal. js writer == realwriter 까지함 . delete 자체는 다시봐야할것.

+++ src/main/resources/static/js/board/modal.js
@@ -22,29 +22,46 @@ $(document).ready(function(){
 		// content 담기
 		var row = $(this).parent().parent().parent();
 		var tr = row.children();
-
 		var title = tr.eq(1).text();
 		var content = tr.eq(3).text();
         var writer = '';
 
+		$("#myModal").modal();
 		$("#modal-title").text("수정 페이지");
-
         $("#title").val(title);
         $("#content").val(content);
 		$("#writer").val(writer);
 
-
-		$("#myModal").modal();
 	});
 
-	// 삭제하기 버튼 클릭
-	$("button[name='delete']").click(function(){
+	//delModal 의 delSubmit 버튼 클릭
+	$("#delSubmit").click(function () {
+		action='delete'
+		url = '/board';
 		bno = this.value;
-		$.ajax({
-			url : '/board/' + bno,
-			type : 'DELETE',
-			complete : function () {location.reload(); }
-		});
+		//alert( $("#writer").val());
+
+		//var writer =  $("input=writer").val();
+		var writer =  $('input#writer')[1].value;
+		var realwriter = $("#realwriter").val();
+
+		console.log(writer);
+
+		if( writer==realwriter) {
+			$.ajax({
+				url: '/board/' + bno,
+				type: 'DELETE',
+				data: data,
+				success: function (data) {
+					$("#delModal").modal('toggle');
+				},
+				complete: function (data) {
+					location.reload();
+				}
+			});
+		}else {
+			alert("입력한비번: "+writer+"진짜비번: "+realwriter);
+		}
 
 	})
 
@@ -74,6 +91,20 @@ $(document).ready(function(){
 
 		})
 	});
+
+	// 삭제하기 버튼 클릭
+	$("button[name='delete']").click(function(){
+		bno = this.value;
+
+		// content 담기
+		var row = $(this).parent().parent().parent();
+		var tr = row.children();
+		var writer = tr.eq(5).text();
+
+		//alert(writer);
+		$("#realwriter").val(writer);
+		$("#delModal").modal();
+	})
 
 
 });
